requirements.txt
Django==5.2.5
pytesseract==0.3.13
openai==1.102.0
google-generativeai==0.8.3
openpyxl==3.1.5
reportlab==4.4.3
python-docx==1.2.0
supabase==2.18.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
PyMuPDF==1.24.10
Pillow==10.4.0
requests==2.32.3